moving rpc shit to own file
